+++ CHANGELOG.md
@@ -51,9 +51,15 @@
 - Added Shutdown-On-SIGINT (CTRL+C shuts down the server)
 
 ## 0.0.7
+
 - Removed Annotation for Flutter and AOT support
 - Better documentation for some methods
 
 ## 0.0.8
+
 - Improved documentation for ALL methods
 - Implemented Hot Reload (1. force hot reload, 2. auto-hot-reload by watching changes)
+- Implemented WebSocket support!
+- Implemented Queue system
+- Implemented Custom Template Engines
+- Added some quite good built-in custom template engines for JSON, YAML, Toml, CSS and Javascript